requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.62.1
kiwisolver==1.5.0
matplotlib==3.10.8
numpy==2.4.3
packaging==26.0
pandas==3.0.1
pillow==12.1.1
py4j==0.10.9.9
pyparsing==3.3.2
pyspark==3.5.8
python-dateutil==2.9.0.post0
pyyaml==6.0.3
ruff==0.15.7
scikit-learn==1.7.2
setuptools==82.0.1
six==1.17.0
